requirements.txt
pytest

langchain-ollama
langchain-community
